+++ src/request-matcher/matchers/query.ts
@@ -5,23 +5,28 @@ import { MockRequestSchema } from '../../protocol';
 import { MatchingContext } from '../context';
 
 export class QueryMatcher {
-  private expectedQuery: Record<string, string | number | null>;
+  private expectedQuery: MockRequestSchema['query'];
+  private noQueryAllowed: boolean;
 
   constructor(private schema: MockRequestSchema) {
-    this.expectedQuery = this.schema.query || {};
+    this.expectedQuery = this.schema.query;
+    // null means the request must not contain any query parameters.
+    this.noQueryAllowed = this.expectedQuery === null;
   }
 
   get hasQuery() {
-    return Object.keys(this.expectedQuery).length > 0;
+    if (this.noQueryAllowed) return true;
+    return Object.keys(this.expectedQuery || {}).length > 0;
   }
 
   match(ctx: MatchingContext) {
-    return Object.keys(this.expectedQuery).every((key) => this.matchQueryParam(ctx, key));
+    if (this.noQueryAllowed) return !ctx.searchParams.toString();
+    return Object.keys(this.expectedQuery || {}).every((key) => this.matchQueryParam(ctx, key));
   }
 
   private matchQueryParam(ctx: MatchingContext, name: string) {
     // todo: handle multi-value params in expectedQuery
-    const expectedValue = this.expectedQuery[name]?.toString() ?? null;
+    const expectedValue = this.expectedQuery?.[name]?.toString() ?? null;
     const actualValues = ctx.searchParams.getAll(name);
 
     const result =